Change led_set_rgb() to accept 16-bit values.
Actual resolution is 10-bit so that we have an acceptably high PWM
frequency (15625 Hz). Resolution limit is transparent to API users. In
addition, duty cycle is scaled to 75% maximum to limit overheating.

Also eliminate unused LED multiplier function. This was a vestige from
when we had an onboard encoder knob.